design : MainPage 디자인 수정
