+++ src/pages/AppRouter.jsx
@@ -8,7 +8,6 @@ import LyricsSelect from './LyricsSelect';
 import Join from './Join';
 import MainPage from './MainPage';
 import OnBoardingPage from './OnBoardingPage';
-import MainPageTest from './MainPageTest';
 
 function AppRouter() {
   return (
@@ -17,7 +16,6 @@ function AppRouter() {
         <Routes>
           <Route path="/" element={<Mypage />} />
           <Route path="/MainPage" element={<MainPage />} />
-          <Route path="/Test" element={<MainPageTest />} />
           <Route path="/search" element={<Search />} />
           <Route path="/lyricsselect" element={<LyricsSelect />} />
           <Route path="/login" element={<Login />} />
